# Improved adler32 NEON performance by 20-30%