We unlocked some ILP by allowing for independent sums in the loop and
reducing these sums outside of the loop. Additionally, the multiplication
by 32 (now 64) is moved outside of this loop.  We have enough register
headroom that we could additionally track some more of the sums in the loop
separately but it's suspected that the widening pairwise additions are the
actual bottlenecks here.  It might be possible with additional branching to do
half of the pairwise additions by summing to 16 bit intermediates but this requires
a fair bit of bookkeeping to get correct. The 16 bit intermediates every 128 loop
iterations need to be summed into 32 bit values to not overflow.

On the Odroid-N2, the Cortex-A73 cores are ~20-25% faster on the adler32 benchmark,
and the Cortex-A53 cores are anywhere from 25-30% faster.