@KungFuJesus 2022-02-12
Comparison for A73:
![a73cmp](https://user-images.githubusercontent.com/3620277/153717602-1bcb04f9-4e41-400d-a102-23566bc5864f.png)

Comparison for A53:
![a53cmp](https://user-images.githubusercontent.com/3620277/153717634-3d65301f-6b1d-4429-8810-7bc58507dd72.png)