@KungFuJesus 2022-02-12
> Wow, nice work again. You must have access to a lot of different CPUs!

Well the Odroid is big.LITTLE, so it has 2 different ARM SIPs on.  But yeah, I have a pretty decent boutique collection, hah.  I don't think what I have is too terribly exotic - the Powermac G5 and the SPARC-T4 may be the more unique of them (though far from _rare_).