@KungFuJesus 2022-02-12
The 16 bit running sums I mentioned earlier in the commit message...I tried that it got kind of messy and slower.  The widening sum to a larger precision appears to be a rather costly instruction (at least for the A72, according to this: https://documentation-service.arm.com/static/5ed75eeeca06a95ce53f93c7?token=)

However, their little footnote confusingly seems to imply that this latency is somehow pipelined away such that the total throughput ends up being 1 per cycle for the accumulation part.  So, what we end possibly losing on the pairwise sums we end up possibly making up for by aggregating the adds in parallel to the 32 bit sums.  Knowing this, we might be able to do at least marginally better by not leveraging the single operand pairwise adds at all.

For the "s2" sums, we come dangerously close to overflowing pretty early with those "tap" constants, but I'm pretty sure I did the math there such that it works out given the where adjacent elements end up landing for the folding part of it.