@KungFuJesus 2022-02-12
Ughh...GCC was annoying late to implementing all of these intrinsics.  Now I have to figure out when they were introduced and add a fallback function, again...