@KungFuJesus 2022-02-13
Well that's really weird - adjusting the n loop iterations that can run during the first round the loop seems to break the fuzzer test and alter the checksums.  I'm doing this to avoid a rebase back to the prime number base of the adler checksum, since it's the maximum number of scalar sums that can be performed with max input bytes before we roll over 32 bits.  

@mtl1979 You mainly did the code for the loop on powerpc, is there something I'm doing that's really wrong?  To make things work I just rebase again, but the commented out bits are still there that conditionally adjust n for when at the beginning of the loop.  The checksums do end up being different but I can't for the life of me figure out why.  Even weirder, I did this same optimization with the VMX code and that doesn't seem to be failing the fuzzer at all (despite it having identical vector widths).