@KungFuJesus 2022-02-13
> @KungFuJesus Main difference between ARM and PowerPC is that ARM is little-endian and PowerPC is big-endian, which can result data getting loaded in different order. My bet is still overflow due to a corner case. With this wide loads few more bytes due to aligning the buffer can be enough to cause overflow even if you try to adjust the first iteration. It all becomes to the question what was the adler32 value when entering the function, if one or both of the components was close to maximum, then it will definitely overflow.
> 
> When adjusting the first iteration, we need to be careful not to round towards zero, as it will result in adjusting being cancelled. Wider the loads, more careful we need to be, as we're basically counting number of remaining iterations, not number of bytes.

I suspect looking for overflow might be the right track. I'll bet I'm overflowing on the 16 bit multiaccumlates since they can't accumulate back to 32 bit values (which is the main difference between this vectorization and all the others). The multiplication is safe but I think that one of the accumulations are biting us. I can probably get around it by modifying where the larger products accumulate back to, though.