@mtl1979 2022-02-13
@KungFuJesus First hunch is often the correct one... Have to eliminate the other possibilities first and the remaining one is the culprit... Like I said before, every time something doesn't work as we planned, we learn more.