@KungFuJesus 2022-02-13
Interesting...it might actually be in the remainder peeling.  I modified the unrolled loop so that it's skipped over and basically does the remainder peeling (with no unroll factor) and I still see the issue.  This is very odd.