@KungFuJesus 2022-02-13
It really doesn't _seem_ like there's any potential overflow in here:
```
    if (rem) {                                                                                   
        adacc_prev = vshlq_n_u32(adacc_prev, 4);                                                 
        while (rem--) {                                                                          
            uint8x16_t d0 = vld1q_u8(buf);                                                       
            uint16x8_t adler, sum2, sum2_lo;                                                     
            adler = vpaddlq_u8(d0);                                                              
            sum2 = vmull_high_u8(t0_t3.val[3], d0);                                              
            sum2_lo = vmull_u8(vget_low_u8(t0_t3.val[3]), vget_low_u8(d0));                      
            adacc = vpadalq_u16(adacc, adler);                                                   
            s2acc = vpadalq_u16(s2acc, sum2);                                                    
            s3acc = vaddq_u32(s3acc, adacc_prev);                                                
            s2acc = vpadalq_u16(s2acc, sum2_lo);                                                 
            adacc_prev = vshlq_n_u32(adacc, 4);                                                  
            buf += 16;                                                                           
        }                                                                                        
    } 
```

That loop is even free of multiply accumulates - yet if I trim off the first 13 iterations in scalar sums and then subtract it from the max loop iterations I get erroneous checksums.