@KungFuJesus 2022-02-13
Hah, bringing in the old version of the inlined function, even, results in incorrect behavior.  Now it may be presumptuous to assume the original implementation was perfectly correct but it does seem like something is up with this sum if I try to trim unaligned starting bytes.

This is really weird.