@mtl1979 2022-02-13
> That loop is even free of multiply accumulates - yet if I trim off the first 13 iterations in scalar sums and then subtract it from the max loop iterations I get erroneous checksums.

That would still imply that the modulo becomes no-op... We basically assume that the modulo is never no-op and that makes sure we never overflow even if we do wide loads.