@mtl1979 2022-02-13
Starting sum should be in range of 0...4294961743... That means before the modulo, the 32-bit sum should be between 4294961743 and 4294967295.