@mtl1979 2022-02-13
> > Starting sum should be in range of 0...4294961743... That means before the modulo, the 32-bit sum should be between 4294961743 and 4294967295.
> 
> I see what you mean but even if I force it to do the first modulo operation anyway and still reduce the max first sums it ends up being wrong, which is really strange.

That would mean it needs to process "tail" bytes before hitting exactly NMAX... Basically the logic assumes the alignment adjust + tail bytes equal 16 bytes.