@Dead2 2022-02-23
RPI 3, aarch64, GCC 10

## Baseline GCC10 321028cde93dac9f47b9b92ad186c4cbffb77cf2
```
 Tool: minideflate Levels: 0-9
 Runs: 32         Trim worst: 18

 Level   Comp   Comptime min/avg/max/stddev  Decomptime min/avg/max/stddev  Compressed size
 0    100.015%      0.020/0.033/0.040/0.005        0.036/0.055/0.064/0.009       15,738,731
 1     54.167%      0.550/0.566/0.576/0.008        0.225/0.248/0.261/0.011        8,523,873
 2     43.877%      0.955/0.966/0.974/0.006        0.231/0.248/0.263/0.011        6,904,637
 3     42.389%      1.291/1.301/1.311/0.006        0.224/0.246/0.257/0.010        6,670,464
 4     41.647%      1.393/1.433/1.444/0.014        0.215/0.239/0.255/0.012        6,553,751
 5     41.226%      1.520/1.540/1.555/0.012        0.220/0.234/0.244/0.009        6,487,426
 6     41.044%      1.772/1.802/1.815/0.012        0.200/0.222/0.238/0.012        6,458,801
 7     40.778%      2.404/2.419/2.430/0.008        0.191/0.227/0.241/0.012        6,416,912
 8     40.703%      2.929/2.949/2.963/0.009        0.212/0.227/0.238/0.007        6,405,232
 9     40.409%      3.309/3.349/3.365/0.014        0.208/0.231/0.249/0.014        6,358,939

 avg1  48.626%                        1.636                          0.218
 avg2  54.028%                        1.817                          0.242
 tot                                228.999                         30.476       76,518,766

   text    data     bss     dec     hex filename
 116222    1664      16  117902   1cc8e libz-ng.so.2
 
----------------------------------------------------------------------
Benchmark                            Time             CPU   Iterations
----------------------------------------------------------------------
adler32/c/2048                    3216 ns         3216 ns       217665
adler32/c/4096                    6404 ns         6404 ns       109217
adler32/c/32768                  48871 ns        48871 ns        14325
adler32/c/262144                392175 ns       392176 ns         1784
adler32/c/2097152              3234376 ns      3233998 ns          216
adler32/c/4194304              6494097 ns      6494141 ns          108
adler32/neon/2048                 1398 ns         1398 ns       500747
adler32/neon/4096                 2744 ns         2743 ns       255071
adler32/neon/32768               22103 ns        22101 ns        31677
adler32/neon/262144             176862 ns       176862 ns         3952
adler32/neon/2097152           1595711 ns      1595351 ns          445
adler32/neon/4194304           3204555 ns      3204200 ns          222
```

## PR #1166 GCC10 50d94bc960980b6dbf54f746dabbc19535a0ff70
```
 Level   Comp   Comptime min/avg/max/stddev  Decomptime min/avg/max/stddev  Compressed size
 0    100.015%      0.016/0.028/0.032/0.005        0.035/0.049/0.060/0.008       15,738,731
 1     54.167%      0.538/0.560/0.573/0.011        0.222/0.244/0.259/0.012        8,523,873
 2     43.877%      0.949/0.969/0.980/0.009        0.212/0.243/0.253/0.010        6,904,637
 3     42.389%      1.253/1.290/1.305/0.016        0.207/0.234/0.248/0.012        6,670,464
 4     41.647%      1.401/1.427/1.443/0.012        0.193/0.228/0.244/0.015        6,553,751
 5     41.226%      1.528/1.550/1.561/0.011        0.211/0.232/0.244/0.011        6,487,426
 6     41.044%      1.781/1.795/1.808/0.009        0.208/0.226/0.236/0.009        6,458,801
 7     40.778%      2.391/2.419/2.432/0.013        0.184/0.216/0.232/0.014        6,416,912
 8     40.703%      2.939/2.960/2.983/0.015        0.198/0.222/0.231/0.010        6,405,232
 9     40.409%      3.305/3.342/3.354/0.014        0.192/0.232/0.248/0.015        6,358,939

 avg1  48.626%                        1.634                          0.213
 avg2  54.028%                        1.816                          0.236
 tot                                228.776                         29.762       76,518,766

   text    data     bss     dec     hex filename
 116734    1664      16  118414   1ce8e libz-ng.so.2

----------------------------------------------------------------------
Benchmark                            Time             CPU   Iterations
----------------------------------------------------------------------
adler32/c/2048                    3216 ns         3216 ns       217723
adler32/c/4096                    6405 ns         6405 ns       109302
adler32/c/32768                  48838 ns        48836 ns        14328
adler32/c/262144                392281 ns       392232 ns         1785
adler32/c/2097152              3244762 ns      3244374 ns          215
adler32/c/4194304              6512863 ns      6512889 ns          108
adler32/neon/2048                  795 ns          795 ns       879910
adler32/neon/4096                 1493 ns         1493 ns       468779
adler32/neon/32768               11984 ns        11984 ns        58389
adler32/neon/262144              99748 ns        99736 ns         7006
adler32/neon/2097152            925089 ns       925041 ns          763
adler32/neon/4194304           1833326 ns      1833336 ns          381
```

Compiled code size increases a bit.
Compression sees no significant change.
Decompression is around 2.3% faster, so that is nice.
Google benchmarks shows nearly double the speed of adler32 neon, very nice.